+++ api/src/utils/quick_utils.ts
@@ -1,646 +1,188 @@
+
+
 import { prisma } from '../database/db';
 import { AquaTree, FileObject, Revision as AquaRevision, OrderRevisionInAquaTree } from 'aqua-js-sdk';
 import { Revision, Link, Signature, WitnessEvent, AquaForms, FileIndex } from '@prisma/client';
 import * as fs from "fs"
-
-// Extend AquaTree interface to include linkedChains
-interface ExtendedAquaTree extends AquaTree {
-    linkedChains?: { [key: string]: ExtendedAquaTree };
-}
-
-// --- Internal Helper Functions --- //
- 
-/**
- * Recursively fetches all ancestor revisions starting from a given hash.
- * @param revisionHash The starting revision hash (including user address prefix).
- * @returns A Promise resolving to an array of Revision objects.
- */
-async function _findAquaTreeRevision(revisionHash: string): Promise<Array<Revision>> {
-    let revisions: Array<Revision> = [];
-
-    // Fetch latest revision 
-    let latestRevionData = await prisma.revision.findFirst({
-        where: {
-            pubkey_hash: revisionHash
-        }
-    });
-
-    if (latestRevionData == null) {
-        // If the initial hash isn't found, log an error or throw
-        console.error(`Unable to get revision with hash ${revisionHash}`);
-        return [];
-    }
-
-    console.log(`Found revision: ${latestRevionData.pubkey_hash}, type: ${latestRevionData.revision_type}`);
-    revisions.push(latestRevionData);
+import { ExtendedAquaTree, ExtendedAquaTreeData } from 'src/models/types';
+import { createAquaTreeFromRevisions } from './revisions_operations_utils';
 
 
-    // Recursively fetch previous revisions if a previous hash exists
-    if (latestRevionData?.previous) {
-        let previousWithPubKey = latestRevionData.previous; // Assume previous already includes pubkey
-
-        // Defensive check: add pubKey prefix if somehow missing
-        if (!previousWithPubKey.includes('_')) {
-            const pubKey = revisionHash.split("_")[0];
-            if (pubKey) { // Ensure pubKey was found
-                previousWithPubKey = `${pubKey}_${latestRevionData.previous}`;
-            } else {
-                console.error(`Could not extract pubKey from ${revisionHash} to prepend to previous hash ${latestRevionData.previous}`);
-                return revisions; // Return what we have so far
-            }
-        }
-
-        // Fetch the ancestor chain recursively
-        try {
-            const ancestorRevisions = await _findAquaTreeRevision(previousWithPubKey);
-            revisions.push(...ancestorRevisions);
-        } catch (error) {
-            console.error(`Error fetching ancestor revision chain for ${previousWithPubKey}:`, error);
-            return revisions; // Return what we have so far
-        }
-    }
-
-    return revisions;
-}
-
-/**
- * Fetches specific data associated with a non-"file" revision type.
- * @param hash The revision hash (including user address prefix).
- * @param revision The Revision object itself.
- * @returns A Promise resolving to the specific data (Signature, WitnessEvent, etc.) or null.
- */
-async function _FetchRevisionInfo(hash: string, revision: Revision): Promise<Signature | WitnessEvent | AquaForms[] | Link | null> {
-    console.log(`Fetching revision info for type: ${revision.revision_type}, hash: ${hash}`);
-
-    switch (revision.revision_type) {
-        case "signature":
-            return await prisma.signature.findFirst({ where: { hash: hash } });
-        case "witness":
-            const witnessLink = await prisma.witness.findFirst({ where: { hash: hash } });
-            if (!witnessLink || !witnessLink.Witness_merkle_root) {
-                console.error(`Witness link or merkle root not found for witness revision ${hash}`);
-                return null; // Or throw error
-            }
-            return await prisma.witnessEvent.findFirst({ where: { Witness_merkle_root: witnessLink.Witness_merkle_root } });
-        case "form":
-            return await prisma.aquaForms.findMany({ where: { hash: hash } });
-        case "link":
-            // Simplify link fetching based on the working implementation
-            console.log(`Querying link table for hash: ${hash}`);
-            const linkData = await prisma.link.findFirst({
-                where: { hash: hash }
-            });
-
-            console.log(`Link data found:`, linkData);
-            return linkData;
-        case "file": // File revisions don't fetch extra info here
-            return null;
-        default:
-            console.warn(`Unknown revision type encountered: ${revision.revision_type} for hash ${hash}`);
-            return null;
-    }
-}
-
-/**
- * Estimates the size in bytes that a string would occupy if saved to a file
- * Uses UTF-8 encoding rules where ASCII chars take 1 byte and others take 2-4 bytes
- * @param str Input string to estimate size for
- * @returns Estimated size in bytes
- */
-function _estimateStringFileSize(str: string): number {
-    if (!str) return 0;
-    return Buffer.byteLength(str, 'utf8'); // Use Buffer for efficiency
-}
-
-// --- Main Public Function --- //
 
 /**
  * Fetches the complete revision chain starting from a given hash,
  * including traversing linked revisions recursively.
- *
- * @param latestHash The starting revision hash (without the user address prefix).
- * @param userAddress The address of the user owning the revision chain.
- * @param url The base URL for constructing file links (currently unused in this version but kept for potential future use).
- * @param _processedHashes Internal set to prevent infinite loops in case of circular links.
- * @param _includeLinkedChains Whether to include full linked chains in the response (default: true).
- * @param _depth Current recursion depth for debugging (default: 0).
- * @param includeFileObjects Whether to include fileObjects in the response (default: false).
- * @returns A Promise resolving to the ExtendedAquaTree containing the complete revision chain and optionally fileObjects.
+ * 
+ * Refactored to use the modular helper functions from the working implementation.
  */
 export async function fetchCompleteRevisionChain(
     latestHash: string,
     userAddress: string,
-    url: string, // Keep url param for potential future file link generation
-    _processedHashes: Set<string> = new Set(), // Add cycle detection
+    url: string,
+    _processedHashes: Set<string> = new Set(),
     _includeLinkedChains: boolean = true,
     _depth: number = 0,
     includeFileObjects: boolean = false
-): Promise<ExtendedAquaTree & { fileObjects?: FileObject[] }> {
+): Promise<ExtendedAquaTreeData> {
     const fullLatestHash = `${userAddress}_${latestHash}`;
-    const indent = "  ".repeat(_depth); // For prettier logging
+    const indent = "  ".repeat(_depth);
 
     console.log(`${indent}[Depth:${_depth}] Fetching chain for ${fullLatestHash}`);
 
-    // Check if we've already processed this hash
+    // Check for circular references
     if (_processedHashes.has(fullLatestHash)) {
         console.warn(`${indent}[Depth:${_depth}] Already processed hash: ${fullLatestHash}. Stopping recursion.`);
-        return { revisions: {}, file_index: {} }; // Return empty tree to prevent infinite recursion
+        return { revisions: {}, file_index: {}, linkedChains: {} };
     }
 
-    // Mark this hash as processed IMMEDIATELY to prevent recursive calls
     _processedHashes.add(fullLatestHash);
 
-    const anAquaTree: ExtendedAquaTree & { fileObjects?: FileObject[] } = {
-        revisions: {},
-        file_index: {},
-        linkedChains: _includeLinkedChains ? {} : undefined
-    };
-    let fileObjects: FileObject[] = []; // Keep track of associated files
-
-    let allRevisionData: Revision[] = [];
     try {
-        console.log(`${indent}[Depth:${_depth}] Fetching chain for ${fullLatestHash}`);
-        // Fetch the main chain starting from the latest hash
-        allRevisionData = await _findAquaTreeRevision(fullLatestHash);
+        // Use the working createAquaTreeFromRevisions function as the base
+        const [baseAquaTree, baseFileObjects] = await createAquaTreeFromRevisions(fullLatestHash, url);
+        
+        // Convert to ExtendedAquaTreeData format
+        let extendedAquaTree: ExtendedAquaTreeData = {
+            revisions: baseAquaTree.revisions,
+            file_index: baseAquaTree.file_index,
+            linkedChains: _includeLinkedChains ? {} : undefined,
+            fileObjects: includeFileObjects ? baseFileObjects : undefined
+        };
 
-        if (allRevisionData.length === 0) {
-            console.warn(`${indent}[Depth:${_depth}] No revisions found for initial hash: ${fullLatestHash}`);
-            return anAquaTree; // Return empty tree if no revisions found
+        // Process linked chains if enabled
+        if (_includeLinkedChains && extendedAquaTree.linkedChains) {
+            extendedAquaTree = await processLinkedChains(
+                extendedAquaTree,
+                userAddress,
+                url,
+                _processedHashes,
+                _depth,
+                includeFileObjects,
+                indent
+            );
         }
 
-        // Also check for "forward" revisions (revisions that reference this one as their previous)
-        // This ensures we get the complete chain including newer revisions
-        // let forwardRevisions: Revision[] = [];
-        // let queryHash = fullLatestHash;
+        console.log(`${indent}[Depth:${_depth}] Successfully processed chain for ${fullLatestHash}`);
+        return extendedAquaTree;
 
-        // while (true) {
-        //     console.log(`${indent}[Depth:${_depth}] Checking for forward revisions from ${queryHash}`);
-        //     const nextRevision = await prisma.revision.findFirst({
-        //         where: {
-        //             previous: queryHash
-        //         }
-        //     });
-
-        //     if (!nextRevision) {
-        //         console.log(`${indent}[Depth:${_depth}] No more forward revisions from ${queryHash}`);
-        //         break;
-        //     }
-
-        //     console.log(`${indent}[Depth:${_depth}] Found forward revision: ${nextRevision.pubkey_hash}, type: ${nextRevision.revision_type}`);
-        //     forwardRevisions.push(nextRevision);
-        //     queryHash = nextRevision.pubkey_hash;
-        // }
-
-        // // If we found forward revisions, add them to the beginning of our list
-        // // so they're processed first (in chronological order)
-        // if (forwardRevisions.length > 0) {
-        //     console.log(`${indent}[Depth:${_depth}] Found ${forwardRevisions.length} forward revisions`);
-        //     allRevisionData = [...forwardRevisions, ...allRevisionData];
-        // }
-
-        // console.log(`${indent}[Depth:${_depth}] Total ${allRevisionData.length} revisions in full chain for ${fullLatestHash}`);
-        // allRevisionData.forEach(rev => {
-        //     console.log(`${indent}[Depth:${_depth}] Chain item: ${rev.pubkey_hash}, type: ${rev.revision_type}`);
-        // });
     } catch (error) {
-        console.error(`${indent}[Depth:${_depth}] Error fetching initial revision chain for ${fullLatestHash}:`, error);
-        throw error; // Re-throw error if initial fetch fails
+        console.error(`${indent}[Depth:${_depth}] Error fetching revision chain for ${fullLatestHash}:`, error);
+        throw error;
     }
+}
 
-    // const result = allRevisionData.reduce((acc: any, item) => {
-    //     acc[item.pubkey_hash] = item;
-    //     return acc;
-    //   }, {});
+/**
+ * Process linked chains recursively using the existing helper functions
+ */
+async function processLinkedChains(
+    aquaTree: ExtendedAquaTreeData,
+    userAddress: string,
+    url: string,
+    processedHashes: Set<string>,
+    depth: number,
+    includeFileObjects: boolean,
+    indent: string
+): Promise<ExtendedAquaTreeData> {
+    const updatedAquaTree = { ...aquaTree };
+    let allFileObjects = [...(aquaTree.fileObjects || [])];
 
-    const revisionsObjects: Record<string, any> = {};
+    // Find all link revisions in the current tree
+    const linkRevisions = Object.entries(aquaTree.revisions).filter(
+        ([_, revision]) => revision.revision_type === 'link'
+    );
 
-    for (const rev of allRevisionData) {
-        revisionsObjects[rev.pubkey_hash] = {
-            ...rev,
-            previous_verification_hash: rev.previous
-        };
-    }
+    for (const [hashOnly, linkRevision] of linkRevisions) {
+        console.log(`${indent}[Depth:${depth}] Processing link revision: ${hashOnly}`);
+        
+        if (linkRevision.link_verification_hashes && linkRevision.link_verification_hashes.length > 0) {
+            for (const linkedHash of linkRevision.link_verification_hashes) {
+                if (typeof linkedHash === 'string' && linkedHash.length > 0) {
+                    console.log(`${indent}[Depth:${depth}] Processing linked hash: ${linkedHash}`);
 
-    // Determine the hash of the earliest revision in this chain (potential genesis)
-    const orderedRevisionData = OrderRevisionInAquaTree({
-        revisions: revisionsObjects,
-        file_index: {}
-    })
-    
-    const orderedRevisionsData = Object.values(orderedRevisionData.revisions)
-    const genesisRevisionInChain = orderedRevisionsData[0];
-    const genesisHashInChain = genesisRevisionInChain.pubkey_hash.split('_')[1];
-
-    // Find associated file indexes based on the genesis hash of this chain
-    let fileIndexes: FileIndex[] = [];
-    if (genesisHashInChain) {
-        try {
-            console.log(`${indent}[Depth:${_depth}] Finding FileIndexes for chain with genesis hash: ${genesisHashInChain}`);
-
-            // Find FileIndex entries where the genesis hash (prefixed) is present in the `hash` array
-            // Note: This logic assumes FileIndex.hash stores prefixed hashes.
-            fileIndexes = await prisma.fileIndex.findMany({
-                where: {
-                    pubkey_hash: {
-                        has: genesisRevisionInChain.pubkey_hash // Check if the prefixed hash exists in the array
+                    // Skip if already processed
+                    const fullLinkedHash = linkedHash.includes('_') ? linkedHash : `${userAddress}_${linkedHash}`;
+                    if (processedHashes.has(fullLinkedHash)) {
+                        console.log(`${indent}[Depth:${depth}] Skipping already processed linked hash: ${linkedHash}`);
+                        continue;
                     }
-                }
-            });
 
-            if (fileIndexes.length > 0) {
-                console.log(`${indent}[Depth:${_depth}] Found ${fileIndexes.length} FileIndexes for genesis hash ${genesisHashInChain}`);
-                for (const fi of fileIndexes) {
-                    console.log(`${indent}[Depth:${_depth}] FileIndex: ${fi.file_hash}`, fi.pubkey_hash);
-                }
-            } else {
-                console.log(`${indent}[Depth:${_depth}] No FileIndexes found for genesis hash, trying non-prefixed hash`);
-throw Error(`No FileIndexes found for genesis hash ${genesisHashInChain}`);
-                // Try with non-prefixed hash as fallback
-                // fileIndexes = await prisma.fileIndex.findMany({
-                //     where: {
-                //         pubkey_hash: {
-                //             has: genesisHashInChain
-                //         }
-                //     }
-                // });
+                    try {
+                        // Update file index for the linked hash using existing helper
+                        const fileIndexValue = await updateFileIndexForLinkedHash(linkedHash);
+                        updatedAquaTree.file_index[linkedHash] = fileIndexValue;
 
-                // if (fileIndexes.length > 0) {
-                //     console.log(`${indent}[Depth:${_depth}] Found ${fileIndexes.length} FileIndexes for non-prefixed genesis hash`);
-                // }
-            }
+                        // Recursively fetch the linked chain
+                        const linkedProcessedHashes = new Set<string>(processedHashes);
+                        const completeLinkedTree = await fetchCompleteRevisionChain(
+                            linkedHash,
+                            userAddress,
+                            url,
+                            linkedProcessedHashes,
+                            true,
+                            depth + 1,
+                            includeFileObjects
+                        );
 
-            // If still no fileIndexes found, try a more general approach for any revision hash
-            if (fileIndexes.length === 0) {
-                console.log(`${indent}[Depth:${_depth}] No FileIndexes found via genesis hash, checking all revisions`);
+                        if (Object.keys(completeLinkedTree.revisions).length > 0) {
+                            console.log(`${indent}[Depth:${depth}] Adding linked tree with ${Object.keys(completeLinkedTree.revisions).length} revisions`);
+                            
+                            // Create compound key for linked chains
+                            const compoundKey = `${hashOnly}_${linkedHash}`;
+                            if (updatedAquaTree.linkedChains) {
+                                updatedAquaTree.linkedChains[compoundKey] = completeLinkedTree;
+                            }
 
-                // Try looking for FileIndex for any revision in this chain
-                for (const rev of allRevisionData) {
-                    const foundIndexes = await prisma.fileIndex.findMany({
-                        where: {
-                            pubkey_hash: {
-                                has: rev.pubkey_hash
+                            // Merge file objects if requested
+                            if (includeFileObjects && completeLinkedTree.fileObjects) {
+                                allFileObjects.push(...completeLinkedTree.fileObjects);
                             }
                         }
-                    });
-
-                    if (foundIndexes.length > 0) {
-                        console.log(`${indent}[Depth:${_depth}] Found ${foundIndexes.length} FileIndexes for revision: ${rev.pubkey_hash}`);
-                        fileIndexes.push(...foundIndexes);
-                        break; // Found some indexes, so stop searching
+                    } catch (linkError) {
+                        console.error(`${indent}[Depth:${depth}] Error processing linked hash ${linkedHash}:`, linkError);
+                        // Set error value for this linked hash
+                        updatedAquaTree.file_index[linkedHash] = '--error--';
                     }
                 }
             }
-
-        } catch (error) {
-            console.error(`${indent}[Depth:${_depth}] Error fetching file indexes for genesis hash ${genesisHashInChain} (full: ${genesisRevisionInChain.pubkey_hash}):`, error);
-            // Decide how to handle: continue without index info, or throw?
         }
     }
 
-
-    // Process each revision in the fetched chain
-    for (const revisionItem of orderedRevisionsData) {
-        console.log(`${indent}[Depth:${_depth}] Revision item: `, revisionItem)
-        const hashOnly = revisionItem.pubkey_hash.split('_')[1];
-        if (!hashOnly) {
-            console.error(`${indent}[Depth:${_depth}] Could not extract hashOnly from pubkey_hash: ${revisionItem.pubkey_hash}`);
-            continue; // Skip this revision if hash is malformed
-        }
-
-        const previousHashOnly = revisionItem.previous
-            ? revisionItem.previous.split('_')[1] ?? ""
-            : "";
-
-        let revisionWithData: AquaRevision = {
-            revision_type: revisionItem.revision_type as AquaRevision['revision_type'],
-            previous_verification_hash: previousHashOnly,
-            local_timestamp: revisionItem.local_timestamp ?? new Date(0).toISOString(), // Use string directly, provide default ISO string
-            version: "https://aqua-protocol.org/docs/v3/schema_2 | SHA256 | Method: scalar", // Example version
-            // Include leaves if they exist
-            ...(revisionItem.verification_leaves && revisionItem.verification_leaves.length > 0 && { leaves: revisionItem.verification_leaves as string[] }),
-        };
-
-        // --- Handle different revision types --- 
-
-        if (revisionItem.revision_type === "file") {
-            // For file revisions, add file_hash and file_nonce
-            revisionWithData.file_nonce = revisionItem.nonce ?? "--error--"; // Use nonce from revision
-            // Find the associated FileIndex to get the definitive file_hash and uri
-            const fileIndexForFileRev = fileIndexes.find(fi => fi.pubkey_hash.includes(revisionItem.pubkey_hash) || fi.file_hash === hashOnly);
-            if (fileIndexForFileRev) {
-                revisionWithData.file_hash = fileIndexForFileRev.file_hash ?? "--error--";
-                // Optionally add the file index URI to the main tree's index
-                if (!anAquaTree.file_index[hashOnly]) { // Add only if not already present (genesis might add it later)
-                    anAquaTree.file_index[hashOnly] = fileIndexForFileRev.file_hash ?? "--error--";
-                }
-                
-                // If we're including fileObjects, create and add a FileObject for this file revision
-                if (includeFileObjects) {
-                    // Create a FileObject based on the file revision information
-                    const fileItem = await prisma.file.findFirst({
-                        where: {
-                            file_hash: fileIndexForFileRev.file_hash,
-                        }
-                    });
-                    const stats = fs.statSync(fileItem!!.file_location!!);
-                    const fileSizeInBytes = stats.size;
-                    const fileObject: FileObject = {
-                        fileName: fileIndexForFileRev.file_hash ?? "--error_uri--",
-                        fileContent: `${url}/files/${fileIndexForFileRev.file_hash}`, // Using file_hash as content reference
-                        path: "./",
-                        // fileSize: fileIndexForFileRev.reference_count ? Number(fileIndexForFileRev.reference_count) : 0 // Using reference_count as a fallback for size
-                        fileSize: fileSizeInBytes
-                    };
-                    
-                    // Add to our fileObjects array
-                    fileObjects.push(fileObject);
-                }
-            } else {
-                console.warn(`${indent}[Depth:${_depth}] FileIndex not found for file revision: ${revisionItem.pubkey_hash}. File hash might be missing.`);
-                revisionWithData.file_hash = "--error_no_index--";
-            }
-        } else {
-            // For non-file types, fetch additional info
-            const revisionInfoData = await _FetchRevisionInfo(revisionItem.pubkey_hash, revisionItem as any);
-
-            if (!revisionInfoData && revisionItem?.revision_type !== 'file' as any) { // file type handled above
-                // Log warning if info is expected but not found
-                console.warn(`${indent}[Depth:${_depth}] Revision info not found for ${revisionItem.revision_type} revision: ${revisionItem.pubkey_hash}`);
-                // Continue processing the revision with available data
-            } else if (revisionInfoData) {
-
-                switch (revisionItem.revision_type) {
-                    case "form":
-                        const formItems = revisionInfoData as AquaForms[];
-                        formItems.forEach(item => {
-                            if (item.key) {
-                                revisionWithData[item.key] = item.value; // Add form key-value pairs
-                            }
-                        });
-                        break;
-                    case "witness":
-                        const witnessData = revisionInfoData as WitnessEvent;
-                        revisionWithData.witness_merkle_root = witnessData.Witness_merkle_root;
-                        revisionWithData.witness_timestamp = witnessData.Witness_timestamp ? parseInt(witnessData.Witness_timestamp, 10) : 0;
-                        revisionWithData.witness_network = witnessData.Witness_network ?? undefined;
-                        revisionWithData.witness_smart_contract_address = witnessData.Witness_smart_contract_address ?? undefined;
-                        revisionWithData.witness_transaction_hash = witnessData.Witness_transaction_hash ?? undefined;
-                        revisionWithData.witness_sender_account_address = witnessData.Witness_sender_account_address ?? undefined;
-                        // revisionWithData.witness_merkle_proof = []; // Merkle proof needs separate handling if required
-                        break;
-                    case "signature":
-                        const signatureData = revisionInfoData as Signature;
-                        try {
-                            // Attempt to parse if it's JSON (e.g., DID signature)
-                            revisionWithData.signature = JSON.parse(signatureData.signature_digest ?? '""');
-                        } catch (e) {
-                            // Otherwise, treat as a plain string
-                            revisionWithData.signature = signatureData.signature_digest ?? "";
-                        }
-                        revisionWithData.signature_public_key = signatureData.signature_public_key ?? undefined;
-                        revisionWithData.signature_wallet_address = signatureData.signature_wallet_address ?? undefined;
-                        revisionWithData.signature_type = signatureData.signature_type ?? undefined;
-                        break;
-                    case "link":
-                        const linkData = revisionInfoData as Link;
-                        revisionWithData.link_type = linkData.link_type ?? undefined;
-                        revisionWithData.link_verification_hashes = linkData.link_verification_hashes as string[];
-                        revisionWithData.link_file_hashes = linkData.link_file_hashes as string[];
-
-                        console.log(`${indent}[Depth:${_depth}] Processing LINK revision: ${revisionItem.pubkey_hash}`);
-                        console.log(`${indent}[Depth:${_depth}] Link verification hashes:`, linkData.link_verification_hashes);
-                        console.log(`${indent}[Depth:${_depth}] Link file hashes:`, linkData.link_file_hashes);
-
-                        // --- Recursively fetch linked chains --- 
-                        if (linkData.link_verification_hashes && linkData.link_verification_hashes.length > 0) {
-                            for (const linkedHash of linkData.link_verification_hashes) {
-                                if (typeof linkedHash === 'string' && linkedHash.length > 0) {
-                                    console.log(`${indent}[Depth:${_depth}] Processing linked hash: ${linkedHash}`);
-
-                                    // Find FileIndex for the linked hash (by pubkey_hash or file_hash)
-                                    let linkedFileIndex = await prisma.fileIndex.findFirst({
-                                        where: { pubkey_hash: { has: linkedHash } }
-                                    });
-                                    if (!linkedFileIndex) {
-                                        linkedFileIndex = await prisma.fileIndex.findFirst({
-                                            where: { file_hash: linkedHash }
-                                        });
-                                    }
-                                    // Find FileName for user-friendly name
-                                    let fileNameEntry = await prisma.fileName.findFirst({ where: { pubkey_hash: linkedHash } });
-                                    anAquaTree.file_index[linkedHash] = fileNameEntry?.file_name ?? linkedFileIndex?.file_hash ?? '--error--';
-
-                                    // For linkedChains feature, we need to recursively fetch each linked chain as a complete tree
-                                    if (_includeLinkedChains && anAquaTree.linkedChains) {
-                                        try {
-                                            // Skip if already processed to prevent infinite recursion
-                                            if (_processedHashes.has(linkedHash)) {
-                                                console.log(`${indent}[Depth:${_depth}] Skipping already processed linked chain: ${linkedHash}`);
-                                                continue;
-                                            }
-
-                                            console.log(`${indent}[Depth:${_depth}] Building complete linked chain for ${linkedHash} to add to linkedChains`);
-
-                                            // First try to find the file index for this linked hash
-                                            // We need this for the genesis revision in the linked chain
-                                            try {
-                                                // Check if we can find a file index for this linked hash
-                                                // This will help populate the file_index for the linked chain's genesis
-
-                                                let pubHash = anAquaTree.linkedChains?.[linkedHash]?.revisions?.[linkedHash]?.link_verification_hashes?.[0] ?? linkedHash
-                                              
-                                                console.log(`${indent}[Depth:${_depth}] Pre-fetching FileIndex for linked chain's genesis: ${pubHash}`);
-                                                const linkedFileIndex = await prisma.fileIndex.findFirst({
-                                                    where: {
-                                                        pubkey_hash: {
-                                                            has: pubHash
-                                                        }
-                                                    }
-                                                });
-
-                                                if (linkedFileIndex) {
-                                                    console.log(`${indent}[Depth:${_depth}] Found FileIndex for linked chain's genesis: ${linkedFileIndex.file_hash}`);
-                                                }
-                                            } catch (indexError) {
-                                                console.warn(`${indent}[Depth:${_depth}] Error pre-fetching FileIndex for linked chain: ${indexError}`);
-                                            }
-
-                                            // Create a new set of processed hashes for this chain to avoid conflicts
-                                            // We use a new Set that INCLUDES all the currently processed hashes
-                                            // to avoid revisiting any already processed revision
-                                            const linkedProcessedHashes = new Set<string>(_processedHashes);
-
-                                            // Fetch the complete chain for this link as a separate tree
-                                            const completeLinkedTree = await fetchCompleteRevisionChain(
-                                                linkedHash,
-                                                userAddress,  // Use the original user address
-                                                url,
-                                                linkedProcessedHashes,
-                                                true, // Include nested chains
-                                                _depth + 1,
-                                                includeFileObjects // Pass down the includeFileObjects parameter
-                                            );
-
-                                            if (Object.keys(completeLinkedTree.revisions).length > 0) {
-                                                console.log(`${indent}[Depth:${_depth}] Adding linked tree to linkedChains with ${Object.keys(completeLinkedTree.revisions).length} revisions`);
-                                                // Store the complete tree in the linkedChains object
-                                                // Extract hashOnly from revisionItem.pubkey_hash
-                                                const parentHashOnly = revisionItem.pubkey_hash.split('_')[1] || '';
-                                                // Create compound key with parent hash and linked hash
-                                                const compoundKey = `${parentHashOnly}_${linkedHash}`;
-                                                // Store with the compound key instead of just linkedHash
-                                                anAquaTree.linkedChains[compoundKey] = completeLinkedTree;
-                                                console.log(`${indent}[Depth:${_depth}] Using compound key: ${compoundKey} for linkedChains`);
-                                                
-                                                // Merge fileObjects from linked chains if they exist
-                                                if (includeFileObjects && completeLinkedTree.fileObjects && completeLinkedTree.fileObjects.length > 0) {
-                                                    fileObjects.push(...completeLinkedTree.fileObjects);
-                                                }
-                                            } else {
-                                                console.warn(`${indent}[Depth:${_depth}] Linked tree for linkedChains is empty: ${linkedHash}`);
-                                            }
-                                        } catch (linkError) {
-                                            console.error(`${indent}[Depth:${_depth}] Error building linked chain for linkedChains: ${linkedHash}`, linkError);
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                        break;
-                } // end switch
-            } // end else if (revisionInfoData)
-        } // end else (non-file type)
-
-        // --- Finalize revision processing --- 
-
-        // Add file index for the genesis revision of the current chain segment
-        if (!previousHashOnly) { // Check if it's a genesis revision (no previous)
-            console.log(`${indent}[Depth:${_depth}] Found genesis revision: ${revisionItem.pubkey_hash}, hash: ${hashOnly}`);
-
-            // Try to find the file index for this genesis hash
-            const genesisFileIndex = fileIndexes.find(item => item.pubkey_hash.includes(revisionItem.pubkey_hash) || item.file_hash === hashOnly);
-
-            if (genesisFileIndex) {
-                console.log(`${indent}[Depth:${_depth}] Found FileIndex for genesis with file_hash: ${genesisFileIndex.file_hash}`);
-                anAquaTree.file_index[hashOnly] = genesisFileIndex.file_hash ?? "--error--";
-                // Add file_hash to the genesis revision itself if it's a file type (might be redundant but ensures presence)
-                if (revisionWithData.revision_type === 'file') {
-                    revisionWithData.file_hash = genesisFileIndex.file_hash ?? "--error--";
-                }
-            } else {
-                // If we didn't find the file index in our initial lookup, try a direct query
-                console.log(`${indent}[Depth:${_depth}] FileIndex not found in pre-fetched indexes, trying direct query for genesis revision: ${revisionItem.pubkey_hash}`);
-
-                // Try to find by exact pubkey_hash first
-                try {
-                    const directFileIndex = await prisma.fileIndex.findFirst({
-                        where: {
-                            pubkey_hash: {
-                                has: revisionItem.pubkey_hash
-                            }
-                        }
-                    });
-
-                    if (directFileIndex) {
-                        console.log(`${indent}[Depth:${_depth}] Found FileIndex for genesis with direct query: ${directFileIndex.file_hash}`);
-                        anAquaTree.file_index[hashOnly] = directFileIndex.file_hash ?? "--error--";
-                        if (revisionWithData.revision_type === 'file') {
-                            revisionWithData.file_hash = directFileIndex.file_hash ?? "--error--";
-                        }
-                    } else {
-                        // Try by a more fuzzy match if needed
-                        const fuzzyFileIndex = await prisma.fileIndex.findFirst({
-                            where: {
-                                pubkey_hash: {
-                                    has: hashOnly
-                                }
-                            }
-                        });
-
-                        if (fuzzyFileIndex) {
-                            console.log(`${indent}[Depth:${_depth}] Found FileIndex for genesis with fuzzy query: ${fuzzyFileIndex.file_hash}`);
-                            anAquaTree.file_index[hashOnly] = fuzzyFileIndex.file_hash ?? "--error--";
-                            if (revisionWithData.revision_type === 'file') {
-                                revisionWithData.file_hash = fuzzyFileIndex.file_hash ?? "--error--";
-                            }
-                        } else {
-                            console.warn(`${indent}[Depth:${_depth}] FileIndex not found for genesis revision: ${revisionItem.pubkey_hash}`);
-                            // Add placeholder URI if index is missing for genesis
-                            anAquaTree.file_index[hashOnly] = "--error_missing_genesis_uri--";
-                            if (revisionWithData.revision_type === 'file') {
-                                revisionWithData.file_hash = "--error_missing_genesis_hash--";
-                            }
-                        }
-                    }
-                } catch (error) {
-                    console.error(`${indent}[Depth:${_depth}] Error querying for genesis FileIndex: ${error}`);
-                    anAquaTree.file_index[hashOnly] = "--error_missing_genesis_uri--";
-                    if (revisionWithData.revision_type === 'file') {
-                        revisionWithData.file_hash = "--error_missing_genesis_hash--";
-                    }
-                }
-            }
-        }
-
-        // Add the processed revision to the tree, avoiding duplicates from linked chains
-        if (!anAquaTree.revisions[hashOnly]) {
-            anAquaTree.revisions[hashOnly] = revisionWithData;
-        } else {
-            // If already present (likely from a link), maybe log or decide if merge logic is needed
-            // console.log(`Revision ${hashOnly} already exists in tree, likely from a link. Skipping add.`);
-        }
-    } // end for loop over allRevisionData
-
-    // Add fileObjects to the result if requested
+    // Update file objects if requested
     if (includeFileObjects) {
-        anAquaTree.fileObjects = fileObjects;
+        updatedAquaTree.fileObjects = allFileObjects;
     }
 
-    // Return the populated AquaTree
-    return anAquaTree;
+    return updatedAquaTree;
 }
 
-// --- Optional Example Usage --- //
-/*
-async function example() {
+/**
+ * Update file index for a linked hash using existing database queries
+ * Returns the file index value instead of mutating the object
+ */
+async function updateFileIndexForLinkedHash(linkedHash: string): Promise<string> {
     try {
-        const latestHash = 'your_latest_revision_hash_here'; // Replace with an actual hash
-        const userAddress = 'user_address_here'; // Replace with the user's address
-        const apiUrl = 'http://localhost:3000'; // Replace with your API URL
-
-        console.log(`Fetching complete chain for ${userAddress}_${latestHash}...`);
-        const completeTree = await fetchCompleteRevisionChain(latestHash, userAddress, apiUrl);
-        console.log('----------------------------------------');
-        console.log('Complete Revision Tree:');
-        console.log(JSON.stringify(completeTree, null, 2));
+        // Use the same logic as the working implementation
+        let linkedFileIndex = await prisma.fileIndex.findFirst({
+            where: { pubkey_hash: { has: linkedHash } }
+        });
         
-        // The result will include:
-        // 1. A "revisions" object with all revisions flattened into one map
-        // 2. A "file_index" mapping verification hashes to URIs
-        // 3. A "linkedChains" object containing complete trees for each linked document
-        //    This creates a nested structure where each linked document has its full chain
-        
-        // To access linked chains:
-        if (completeTree.linkedChains) {
-            for (const hashOnly in completeTree.linkedChains) {
-                const linkedTree = completeTree.linkedChains[hashOnly];
-                console.log(`Linked chain with hash ${hashOnly} has ${Object.keys(linkedTree.revisions).length} revisions`);
-                
-                // Each linked tree may itself have more linked chains (nested structure)
-                if (linkedTree.linkedChains) {
-                    console.log(`This linked chain also has ${Object.keys(linkedTree.linkedChains).length} of its own linked chains`);
-                }
-            }
+        if (!linkedFileIndex) {
+            linkedFileIndex = await prisma.fileIndex.findFirst({
+                where: { file_hash: linkedHash }
+            });
         }
-        
-        console.log('----------------------------------------');
+
+        // Find FileName for user-friendly name
+        let fileNameEntry = await prisma.fileName.findFirst({ 
+            where: { pubkey_hash: linkedHash } 
+        });
+
+        return fileNameEntry?.file_name ?? linkedFileIndex?.file_hash ?? '--error--';
     } catch (error) {
-        console.error('Error in example usage:', error);
+        console.error(`Error updating file index for linked hash ${linkedHash}:`, error);
+        return '--error--';
     }
 }
-
-// example(); // Uncomment to run the example
-*/
 
 /**
  * Diagnostic function to check link table structure and data
